Allowed accelerator to enter doubles on product search

+++ apps/purchases/general/product/product.jsx
@@ -38,7 +38,7 @@ export default class Product extends React.Component {
       if (ev.target.value) {
         let bits = ev.target.value.split('*')// Split val [0] is code [1] is qty
         const code = bits[0]
-        const qty = isNaN(bits[1]) ? 1 : parseInt(bits[1])
+        const qty = isNaN(bits[1]) ? 1 : parseFloat(bits[1])
         this.props.dispatch(searchProduct(code, 'product', 'productSearch', qty, this.props.itemsInCart))
         this.props.dispatch({type: 'CLEAR_PRODUCT_FIELD_VALUE'})
